==> scripts/7-GLAD_export.ipynb <==
### 1.0 Libraries and Directories

import ee
import geemap
import pprint as pp

ee.Authenticate()
ee.Initialize()

glad = ee.ImageCollection('projects/glad/water/individualMonths')

rois = ee.data.listAssets('projects/alpod-412314/assets/ROIs/')
rois_ids = [roi['id'] for roi in rois['assets']]
rois_ids.remove('projects/alpod-412314/assets/ROIs/all_regions')
pp.pp(rois_ids)

### 2.0 Combine all the GLAD observations for June & August for 20 years. 

years = range(1990, 2025) # 2016-2020 is the common timeframe | 1999-2020 is the full timeframe

aug_sys_idx =[f'{year}_08' for year in years]
june_sys_idx =[f'{year}_06' for year in years]

jun_imgs = glad.filter(ee.Filter.inList('system:index', june_sys_idx))
jun_max_obs = jun_imgs.size().getInfo()
aug_imgs = glad.filter(ee.Filter.inList('system:index', aug_sys_idx))
aug_max_obs = aug_imgs.size().getInfo()

monthly_datasets = [
    (jun_imgs, jun_max_obs, 'june'), 
    (aug_imgs, aug_max_obs, 'aug'),
]

print(jun_max_obs, aug_max_obs)

pp.pp(jun_imgs.getInfo())

def select_invalid(img, mask_val):
    img_unmasked = img.unmask(mask_val)
    inval_pixels = img_unmasked.eq(mask_val)
    inval_pixels = inval_pixels.updateMask(ee.Image.constant(1))
    return inval_pixels

def mask_invalid(img, mask_val):
    data_mask = img.neq(mask_val)
    data = img.updateMask(data_mask)
    return data

def calc_valid_occurance(img_collection, max_obs, mask_val):
    # Sum the invalid pixels in each image
    invalid_masks = img_collection.map(lambda img: select_invalid(img, mask_val))
    sum_invalid = invalid_masks.reduce(ee.Reducer.sum())

    # Sum the valid water observations in each image
    data = img_collection.map(lambda img: mask_invalid(img, mask_val))
    sum_water = data.reduce(ee.Reducer.sum())

    # Calculate the % valid occurence
    max_obs_img = ee.Image.constant(max_obs)
    denom_img = max_obs_img.subtract(sum_invalid)
    valid_occurence = sum_water.divide(denom_img)

    return valid_occurence



target_imgs = []

for ds in monthly_datasets:
    valid_occurance = calc_valid_occurance(ds[0], ds[1], 255)
    month = ds[2]
    target_imgs.append((valid_occurance, month))

june = target_imgs[0][0]


Map = geemap.Map()
viz_params = {
    'min': 0,
    'max': 22,  
    'palette': ['red', 'yellow', 'blue']
}
Map.addLayer(june, viz_params, 'Summed Image')

Map

def export_rois(roi_id, target_img, month):
    
    roi_name_temp = roi_id.split('/')[-1]
    roi_name = roi_name_temp.split('_roi')[0]
    feature = ee.FeatureCollection(roi_id).first()
    bounds = feature.geometry()

    clipped_img = target_img.clip(bounds)

    export_params = {
        'image': clipped_img,
        'description': f'gladFullMonthly_{roi_name}_month_{month}_', 
        'fileNamePrefix': f'gladFullMonthly_{roi_name}_month_{month}_',
        'folder': 'glad_exports', 
        'scale': 30,
        'crs': 'EPSG:4326',
        'region': bounds,
        'fileFormat': 'GeoTIFF',
        'maxPixels': 1e13
    }

    ee.batch.Export.image.toDrive(**export_params).start()

print(target_imgs)

for roi_id in rois_ids:
    for img in target_imgs:
        image = img[0]
        month = img[1]
        export_rois(roi_id, image, month)
